--- components/portfolio/InventoryItemFilter.module.css
@@ -1,3 +1,0 @@
-.container {
-  @apply flex flex-col border-1 border-black min-w-[290px] w-[290px] max-w-[290px] p-10;
-}
