+++ components/portfolio/inventory/item/InventoryItemFilter.module.css
@@ -1,0 +1,3 @@
+.container {
+  @apply flex flex-col border-1 border-black min-w-[290px] w-[290px] max-w-[290px] p-10 max-h-[400px];
+}
