+++ components/portfolio/inventory/item/InventoryItemList.module.css
