+++ components/portfolio/inventory/item/InventoryItemSection.tsx
